requirements.txt
Django==5.2.5
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.7.0
django-filter==25.1
drf-yasg==1.21.10
Pillow==11.3.0
python-decouple==3.8
PyJWT==2.10.1
cryptography==43.0.0
